reorganizing some files
